Update it.md
Line 339: I'd avoid the use of 'password' close to 'passphrase'. The majority of new comers confuse them when dealing with passphrase. A Wallet passoword is not a passphrase. This lead them to load different wallet with (of course) 0 balance. 
I'd personally change the same on en.md line 336. Let me know. 

Line 587: I think we should teach people to use official resources as much as possibile.So i changed the link accordingly.
Plus, the link provided list words in Hexadecimal notation that i suppose might be hard to deal with.