Update rn.md
Today, we have reached a very important milestone by beginning the proofreading process. One key observation is that many words have been translated too literally, which often results in the loss of the overall meaning in some parts of the text. However, there are other sections that are quite understandable and require only minor adjustments. Another point to note is the presence of certain vocabulary borrowed from Kinyarwanda, which will need to be retranslated into Kirundi