Update rn.md
After reading the entire document today, I have just begun the translation to complete the part I had already started. The observation remains the same: the translation is generally done, but there are parts where it is a literal translation and others few where the meaning is lost. Another issue is that some words are translated into Kinyarwanda. However, with the improvements we are going to make, we will ensure that it is understandable and that it conveys the correct meaning, which is the meaning of the original text in English