proofread es.md
A couple of details... Bitcoin Core's latest version is now 2.9.0 so may be is a good idea to update the info. It looks like part of the info was taken from "Agora256" and its name is still on the tutorial. There is no the "leave us a green thumb up" final sentence but a "feel free to share them with the Agora256 community."