Update rn.md
There are terms that are not available in Kirundi, so we borrow these terms from Kinyarwanda. This remains very understandable because the two languages are almost similar, and almost the entire Burundian population understands Kinyarwanda. Borrowing certain terms from Kinyarwanda therefore does not hinder comprehension in Kirundi. The terms amahera and amafaranga mean the same thing. I have also tried to put some terms in parentheses to facilitate understanding of the meaning of certain words.