Update de.md - 2.2
After "**Funktion $\Sigma_1$:**", there is a problem with formatting - because of the wrongly placed dollar sign, I think. Also, the images and the rest of the text is missing afterwards - I guess, this is fixed by inserting the new line after the dollar sign but please double check. Nevertheless, I proofread the rest of the section from the sources. ;)